This is a somewhat experimental change that attempts to make the doctests easier to understand by displaying relevant snippets of code from the child frame.
I'm not sure how effective it is, but it does help me understand how to use jschannel a bit better by juxtaposing child and parent code on the same page.

This enhancement is purely for the sake of readability, however, and shouldn't impact the execution of tests in any way.